Move the collectionContrib project to a collection-contrib subdirectory
This is a tentative to address #51

At least now, the `package` task builds a .jar file that contains .class files. The `publishLocal` task, however, still does not build a .jar file containing the .class files.

I’ve noticed that `collectionContribJS/publishLocal` builds a correct .jar file. The only difference between the JS and JVM projects is that the JVM project applies the `scalaModuleSettingsJVM` settings, which does something with sbt-osgi, which I didn’t dig into…